backend/requirements.txt
aiosmtplib==3.0.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
attrs==25.3.0
bcrypt==4.0.1
blinker==1.9.0
cachetools==6.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
choreographer==1.0.9
click==8.2.1
contourpy==1.3.2
cryptography==45.0.6
cycler==0.12.1
Deprecated==1.2.18
diskcache==5.6.3
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.14
fastapi-mail==1.5.0
fonttools==4.58.5
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kaleido==1.0.0
kiwisolver==1.4.8
limits==5.5.0
logistro==1.1.0
MarkupSafe==3.0.2
matplotlib==3.10.3
narwhals==2.0.0
numpy==2.3.1
orjson==3.11.1
packaging==25.0
pandas==2.3.1
passlib==1.7.4
pillow==11.3.0
plotly==6.2.0
pluggy==1.6.0
protobuf==6.31.1
pyarrow==21.0.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
PyMySQL==1.1.1
pyparsing==3.2.3
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
reportlab==4.4.2
requests==2.32.4
rpds-py==0.26.0
rsa==4.9.1
simplejson==3.20.1
six==1.17.0
slowapi==0.1.9
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.2
streamlit==1.47.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
wrapt==1.17.3
